+++ tests/integrate_openapi/request_and_validate_test.py
@@ -1,59 +1,15 @@
 """Test client.utils.request_and_validate module."""
 
-import socket
-from typing import AsyncIterator, Callable
+from typing import Callable
 
 import openapi_core
 import pytest
-import pytest_asyncio
 import requests
-import tornado
 from jsonschema_path import SchemaPath
 from openapi_core.validation.response import exceptions
-from tornado.web import RequestHandler
 
 from rest_tools.client import RestClient
 from rest_tools.client.utils import request_and_validate
-from rest_tools.server import RestServer, RestHandlerSetup
-
-
-@pytest.fixture
-def port() -> int:
-    """Get an ephemeral port number."""
-    # unix.stackexchange.com/a/132524
-    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    s.bind(("", 0))
-    addr = s.getsockname()
-    ephemeral_port = addr[1]
-    s.close()
-    return ephemeral_port
-
-
-@pytest_asyncio.fixture
-async def server(port: int) -> AsyncIterator[Callable[[], RestClient]]:
-    """Start up REST server and attach handlers."""
-
-    class TestHandler(RequestHandler):
-        ROUTE = "/echo/this"
-
-        def post(self) -> None:
-            if self.get_argument("raise", None):
-                raise tornado.web.HTTPError(400, self.get_argument("raise"))
-            self.write(self.get_argument("echo"))
-
-    args = RestHandlerSetup({"debug": True})
-    rs = RestServer(debug=True)
-    rs.add_route(TestHandler.ROUTE, TestHandler, args)
-    rs.startup(address="localhost", port=port)
-
-    def client() -> RestClient:
-        return RestClient(f"http://localhost:{port}", retries=0)
-
-    try:
-        yield client
-    finally:
-        await rs.stop()  # type: ignore[no-untyped-call]
-
 
 OPENAPI_SPEC = openapi_core.OpenAPI(
     SchemaPath.from_dict(
